+++ test/truffle/compiler/stf-optimises.sh
@@ -2,4 +2,4 @@
 
 source test/truffle/common.sh.inc
 
-jt ruby --graal --jvm.Dgraal.TruffleCompilationExceptionsAreFatal=true test/truffle/compiler/stf-optimises/stf-optimises.rb
+jt ruby --graal --vm.Dgraal.TruffleCompilationExceptionsAreFatal=true test/truffle/compiler/stf-optimises/stf-optimises.rb