+++ tool/jt.rb
@@ -26,7 +26,7 @@ PROFILES_DIR = "#{TRUFFLERUBY_DIR}/profiles"
 TRUFFLERUBY_GEM_TEST_PACK_VERSION = "8b57f6022f0fa17ace7c8d2a3af730357715e0a2"
 
 JDEBUG_PORT = 51819
-JDEBUG = "--jvm.agentlib:jdwp=transport=dt_socket,server=y,address=#{JDEBUG_PORT},suspend=y"
+JDEBUG = "--vm.agentlib:jdwp=transport=dt_socket,server=y,address=#{JDEBUG_PORT},suspend=y"
 JEXCEPTION = "--exceptions.print_uncaught_java=true"
 METRICS_REPS = Integer(ENV["TRUFFLERUBY_METRICS_REPS"] || 10)
 DEFAULT_PROFILE_OPTIONS = %w[--cpusampler --cpusampler.SampleInternal=true --cpusampler.Mode=roots --cpusampler.Output=json]
@@ -192,7 +192,7 @@ module Utilities
     else
       raise 'set one of GRAALVM_BIN or GRAAL_HOME in order to use Graal'
     end
-    [javacmd, vm_args.map { |arg| "--jvm.#{arg[1..-1]}" } + options]
+    [javacmd, vm_args.map { |arg| "--vm.#{arg[1..-1]}" } + options]
   end
 
   def find_auto_graal_home
@@ -717,12 +717,12 @@ module Commands
         graal = true
       when '--stress'
         graal = true
-        vm_args << '--jvm.Dgraal.TruffleCompileImmediately=true'
-        vm_args << '--jvm.Dgraal.TruffleBackgroundCompilation=false'
-        vm_args << '--jvm.Dgraal.TruffleCompilationExceptionsAreFatal=true'
+        vm_args << '--vm.Dgraal.TruffleCompileImmediately=true'
+        vm_args << '--vm.Dgraal.TruffleBackgroundCompilation=false'
+        vm_args << '--vm.Dgraal.TruffleCompilationExceptionsAreFatal=true'
       when '--asm'
         graal = true
-        vm_args += %w[--jvm.XX:+UnlockDiagnosticVMOptions --jvm.XX:CompileCommand=print,*::callRoot]
+        vm_args += %w[--vm.XX:+UnlockDiagnosticVMOptions --vm.XX:CompileCommand=print,*::callRoot]
       when '--jdebug'
         vm_args << JDEBUG
       when '--jexception', '--jexceptions'
@@ -730,18 +730,18 @@ module Commands
       when '--server'
         vm_args += %w[--instrumentation_server_port=8080]
       when '--infopoints'
-        vm_args << "--jvm.XX:+UnlockDiagnosticVMOptions" << "--jvm.XX:+DebugNonSafepoints"
-        vm_args << "--jvm.Dgraal.TruffleEnableInfopoints=true"
+        vm_args << "--vm.XX:+UnlockDiagnosticVMOptions" << "--vm.XX:+DebugNonSafepoints"
+        vm_args << "--vm.Dgraal.TruffleEnableInfopoints=true"
       when '--fg'
-        vm_args << "--jvm.Dgraal.TruffleBackgroundCompilation=false"
+        vm_args << "--vm.Dgraal.TruffleBackgroundCompilation=false"
       when '--trace'
         graal = true
-        vm_args << "--jvm.Dgraal.TraceTruffleCompilation=true"
+        vm_args << "--vm.Dgraal.TraceTruffleCompilation=true"
       when '--igv', '--igv-full'
         graal = true
-        vm_args << (arg == '--igv-full' ? "--jvm.Dgraal.Dump=:2" : "--jvm.Dgraal.Dump=TruffleTree,PartialEscape:2")
-        vm_args << "--jvm.Dgraal.PrintGraphFile=true" unless igv_running?
-        vm_args << "--jvm.Dgraal.PrintBackendCFG=false"
+        vm_args << (arg == '--igv-full' ? "--vm.Dgraal.Dump=:2" : "--vm.Dgraal.Dump=TruffleTree,PartialEscape:2")
+        vm_args << "--vm.Dgraal.PrintGraphFile=true" unless igv_running?
+        vm_args << "--vm.Dgraal.PrintBackendCFG=false"
       when '--no-print-cmd'
         options[:no_print_cmd] = true
       when '--exec'
@@ -1038,7 +1038,7 @@ module Commands
 
     truffle_args = []
     if !extra_args.include?('--native')
-      truffle_args += %w[--jvm.Xmx2G --jvm.ea --jvm.esa --jexceptions]
+      truffle_args += %w[--vm.Xmx2G --vm.ea --vm.esa --jexceptions]
     end
 
     env_vars = {
@@ -1384,7 +1384,7 @@ EOS
         verify_native_bin!
         options << '-t' << find_launcher(true)
       else
-        options += %w[-T--jvm.ea -T--jvm.esa -T--jvm.Xmx2G]
+        options += %w[-T--vm.ea -T--vm.esa -T--vm.Xmx2G]
         options << "-T--core.load_path=#{TRUFFLERUBY_DIR}/src/main/ruby"
       end
 
@@ -1529,7 +1529,7 @@ EOS
     samples = []
     METRICS_REPS.times do
       log '.', "sampling\n"
-      out = run_ruby '--jvm.Dtruffleruby.metrics.memory_used_on_exit=true', '--jvm.verbose:gc', *args, capture: true, :err => :out, no_print_cmd: true
+      out = run_ruby '--vm.Dtruffleruby.metrics.memory_used_on_exit=true', '--vm.verbose:gc', *args, capture: true, :err => :out, no_print_cmd: true
       samples.push memory_allocated(out)
     end
     log "\n", nil
@@ -1603,7 +1603,7 @@ EOS
   end
 
   def can_run_in_heap(heap, *command)
-    run_ruby("--jvm.Xmx#{heap}M", *command, err: '/dev/null', out: '/dev/null', no_print_cmd: true, continue_on_failure: true, timeout: 60)
+    run_ruby("--vm.Xmx#{heap}M", *command, err: '/dev/null', out: '/dev/null', no_print_cmd: true, continue_on_failure: true, timeout: 60)
   end
 
   def metrics_maxrss(*args)
@@ -1677,8 +1677,8 @@ EOS
 
   def metrics_time_measure(use_json, *args)
     native = args.include? '--native'
-    metrics_time_option = "#{native ? '--native.D' : '--jvm.D'}truffleruby.metrics.time=true"
-    verbose_gc_flag = native ? '--native.XX:+PrintGC' : '--jvm.verbose:gc' unless use_json
+    metrics_time_option = '--vm.Dtruffleruby.metrics.time=true'
+    verbose_gc_flag = native ? '--vm.XX:+PrintGC' : '--vm.verbose:gc' unless use_json
     args = [metrics_time_option, *verbose_gc_flag, '--no-core-load-path', *args]
 
     samples = METRICS_REPS.times.map do
@@ -1826,7 +1826,7 @@ EOS
 
     unless benchmark_ruby
       run_args.push '--graal' unless args.delete('--no-graal') || args.include?('list')
-      run_args.push '--jvm.Dgraal.TruffleCompilationExceptionsAreFatal=true'
+      run_args.push '--vm.Dgraal.TruffleCompilationExceptionsAreFatal=true'
     end
 
     run_args.push "-I#{find_gem('benchmark-ips')}/lib" rescue nil
@@ -2414,13 +2414,7 @@ EOS
       end
 
       configs.each do |config|
-        if config == '--jvm'
-          d = '--jvm.'
-        else
-          d = '--native.'
-        end
-
-        lines.push "RUN " + setup_env["ruby #{config} #{d}Dgraal.TruffleCompilationExceptionsAreThrown=true #{d}Dgraal.TruffleIterativePartialEscape=true --basic_ops.inline=false pe/pe.rb"]
+        lines.push "RUN " + setup_env["ruby #{config} --vm.Dgraal.TruffleCompilationExceptionsAreThrown=true --vm.Dgraal.TruffleIterativePartialEscape=true --basic_ops.inline=false pe/pe.rb"]
       end
     end